Libft Functions

+++ Libft/Source/ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: gamaral <gamaral@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/21 22:17:56 by gamaral           #+#    #+#             */
-/*   Updated: 2023/04/27 22:44:44 by gamaral          ###   ########.fr       */
+/*   Updated: 2023/05/06 20:00:42 by gamaral          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,12 +15,12 @@
 char	*ft_strjoin(char const *s1, char const *s2)
 {
 	char	*aux;
-	int		i;
-	int		j;
+	size_t	i;
+	size_t	j;
 
 	aux = malloc(sizeof(char) * (ft_strlen(s1) + ft_strlen(s2)) + 1);
 	if (!aux)
-		return(NULL);
+		return (NULL);
 	i = 0;
 	j = 0;
 	while (i < ft_strlen(s1))
